requirements.txt
streamlit
pandas
requests
plotly
reportlab
python-pptx
fastapi
uvicorn[standard]
python-dotenv
google-generativeai
markdown
watchdog
python-docx
scipy
scikit-learn
groq
